\docType{data}
\name{NPRI}
\alias{NPRI}
\title{
National Pollutant Release Inventory (NPRI) for Canada, 1993-2012
}
\description{
A dataset containing NPRI entries for the period of 1993 to 2012. 
}
\usage{
load(NPRI)
}
\format{
A data frame with 470641 observations on 81 variables.
}
\source{
The data were obtained from Environment Canada.
}
\examples{
head(NPRI)
}
